+++ jeap-spring-boot-web-config-starter/src/main/java/ch/admin/bit/jeap/web/configuration/servlet/ServletWebConfiguration.java
@@ -26,10 +26,13 @@ public class ServletWebConfiguration {
      * See {@link org.springframework.web.filter.ShallowEtagHeaderFilter} for details. The Etag header avoids sending
      * a response when it is already cached on the client. As such, it saves on bandwidth but not on server-side
      * processing time. The instantiation of the filter bean can be disabled with the property jeap.web.headers.etag=false.
+     * We created our own subclass of the {@link org.springframework.web.filter.ShallowEtagHeaderFilter} to exlcude
+     * Server-Sent Events (SSE) specifically. Without it the SSEs were waiting infinitly until the response is complete,
+     * which never happens because it's an infinite stream.
      */
     @Bean
     @ConditionalOnProperty(name = "jeap.web.headers.etag", matchIfMissing = true)
     ShallowEtagHeaderFilter shallowEtagHeaderFilter() {
-        return new ShallowEtagHeaderFilter();
+        return new SseAwareEtagHeaderFilter();
     }
 }