+++ CHANGELOG.md
@@ -8,6 +8,16 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 > - Spring Boot 3 maintenance (bug fixes, patches, and regular updates) continues on branch `release/springboot3`.
 > - The Spring Boot 4 upgrade process happens on the `master` branch, with the goal of releasing Spring Boot 4 compatible versions of all jeap components in a single major release.
 
+## [22.5.1-alpha-springboot4] - 2026-05-15
+
+### Added
+- Custom subclass of `ShallowEtagFilter` called `SseAwareEtagHeaderFilter`
+  It ignores requests with Accept: text/event-stream as the `ShallowEtagFilter` was waiting for the response to complete
+  which never happens as it is a stream.
+
+### Changed
+- Set `SseAwareEtagHeaderFilter` as default bean for the ShallowEtagFilter. 
+
 ## [22.5.0-alpha-springboot4] - 2026-05-07
 
 ### Added